backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
sqlalchemy
pydantic
